Reversed order so the frontend lexer initializes the backend lexer.
This ends the backward feeling that things were getting. fs_main.c now
initializes the fe parser + lexer. The front and backend scanners still share
data with each other.

fs_main.c now defines functions as lily_impl_* instead of lily_be_*. This will
be the prefix for functions to be defined by the implementation.

For now, lily_impl_fatal calls C's exit, but that is expected to change later.