Merge lily_page_scanner.c code into lily_lexer.c
read_page_line would have needed to be copied into the lexer eventually, so
that was going to be duplicate work. Apart from that, about 2/3rds of the code
size was in syncing data for a func that was less than 50 lines.

lily_page_scanner.{c,h} are now dead, as is the backend directory...I hope.

The Makefile was updated for the change.