Add types for int, float, and list. Implement assign in the vm.
The vm was done first in hopes that it would cause the lower levels to be more
sane. This may or may not work.

This code also carries the implicit assumption that someone besides the vm
will be responsible for creating the list from the temp in a symbol. It has
not been decided what will be responsible for creating the list though.