Make a pool for ASTs and AST lists.
lily_ast_pool type created which holds a pool of N trees, and N lists, with N
given at the start and expanding as necessary. This fixes memory used by ASTs
being lost.

Hopefully, it will be a performance improvement by reducing what would be
large numbers of common malloc and free calls, but it's too early to tell
that.