Handle numbers with decimals, hopefully.
They are scanned in as doubles, instead of floats for more flexibility.
Hopefully the backend will work.

The scanner only supports 8.8 in terms of size, even though doubles might be
able to hold more depending on the platform. That should work, for now.

Now, the only non-working test in basic_types.ly is the list assignment.