Remove lily_impl_{malloc,realloc,free} functions in favor of defines.
Support for allocation implementations will eventually be added back in (in
particular, for helping to support per-thread allocs for per-thread
interpreters).

This was also done so that per-interpreter alloc failures could raise
per-interpreter exceptions. Something impossible without sending the
interpreter for each malloc/realloc/(free?). Additionally, such a method would
make it impossible to free function-local memory not yet placed into the
interpreter somewhere.

For now, allocations are not checked for failure. That will be fixed when
per-interpreter exceptions come.