Interpreter divided into {lex,parse,emit} states, symtab, and error object.
There are a couple advantages to this:
* Reduces the number of unnecessary includes (including the symtab caused the
  lexer to be included, for example).
* The interpreter now has access to all data so it can be properly freed on
  closure (soon!)

When allocating initial data, the lexer, emitter, symtab, and parser return
NULL on allocation failures.

The symtab keeps track of the lexer's line number, so that doesn't have to be
sent with every call for a new symbol. Finally, each of the structures has a
pointer to the new error object so they can raise an exception.