Introduce lily_shutdown, which frees interpreter memory.
Valgrind shows a leak on both tests, due to strings not being freed. This is
the result of a deeper problem (not being able to tell if a string is a
reference to another, or the one being referred to).

Some reference counting is needed before those leaks can be fixed.