Added classes. Declarations require class names. Allow declaration lists.
The idea was to be typed from the beginning, but that was always put behind in
favor of other things. Classes do not support methods just yet, but the basic
structure exists.

Declarations are the standard fare of:
integer a
integer b, c = 1, d

The 'basic_types.ly' test was updated for this, and it supposedly works. The
print function does not check the types though.

Comma was added to the lexer for this.

Initial classes are: function, str, integer (C int), and number (C double).