Added lily_value to move values from lexer to parser to symtab.
Fixed a redundant final assignment in add_symbol.

Previously, there were three entry points for creating a nameless symbol in
the symtab. This allowed combining them into lily_new_temp.

But more importantly, the lexer's token is now the actual token, and values
are stored in lexer->value. The lexer now creates the lily_strval when it gets
a string, instead of packing it in word_buffer.

word_buffer was changed to label, and is now only used for variable names.