Use fixed values and storages instead of registers.
Fixed values are values that are collected by the lexer, and will not change.
Storages are values that exist to hold temporary results (like the result from
some addition). Storages are not created or used yet.

The symtab dumper now categorizes classes, fixed values, storages, and
symbols.

The lexer allocates a lily_strval when it gets a string, instead of reusing
the buffer used for labels.

The removal of registers was part of an experiment to see if there would be
notable performance gains by using addresses directly for ops.