ASTs store func args in arg_start, arg_top. Var values go to result now.
The first change fixes function arguments being run out of order. AST
num_args changed to args_collected, which is more accurate. next_arg added to
ASTs to handle the new style of function arg collection. This had the added
bonus of being able to remove the ast_list type.

Whenever a var is loaded, the value is saved to the AST result instead of
data.object. This allows the emitter to do nothing when it encounters a var
(instead of assigning data.object to result). So, simplification!