Add method support, and opcodes for plus.
Addition currently doesn't work, but this is a big step forward. Methods are
extremely basic right now (only intended to support binary ops, vm opcodes).
But it's a starting ground for actual methods in the future.

The lexer, parser, and vm don't support plus though. So it doesn't work yet.