Add storage symbols. The emitter now counts different expressions.
Storage symbols are used to hold intermediate values, so they're necessary for
things like plus. All classes have at least one storage allocated to them, so
the emitter doesn't have to check for a NULL storage (something that would be
false 99% of the time).

Each expression has a different expr_num in the emitter. This will be used to
make sure that the same storage is not used twice in a single expression. It
will be necessary later for functions where the arguments are expressions
(since each arg will use a different storage).

This is a step toward plus working.