Drop the 'exception' error code.
This is an awful hack for exception support which probably won't exist for a
long time. I'm not even sure if this would be useful to the caller, since most
errors would fall under a 'syntax error' group or 'out of memory'. So it
wouldn't be much help anyway.

It can be added back if it's found to be useful.