Add line_adjust to error object. Emitter records line adjusts.
Suppose there's an expression such as : 'integer a = 1.0 +
1.0'

The lexer will have line 2, but this should really report line 1 (where the
assignment is) for simplicity. So if error's line_adjust is not 0, then it
should be used for showing the line number. So far, this is just for the
emitter.