Remove 'data' field from ASTs by bringing everything in it up.
This increases memory usage by about 12 bytes per struct. The fields which are
supposed to be conditionally accessed are marked appropriately in lily_ast.h

This change was done to simplify code using ast. The worst offender was trying
to get a tree parent's priority:

ast->data.bin_expr.parent->data.bin_expr.priority

is now

ast->parent->priority

(This is the worst example, and was the motivation for the change).