Syms now use signatures instead of classes.
Signatures are better, because they hold the class, but can hold extra in the
case of functions. So it's a merging of class info and properties in one area.

This will result in comparisons of types being faster for the emitter, since
it won't have to check for class and property matches. Signatures can be
shared too, reducing the memory usage. No special cases for extra properties,
for now.

There's no solid reason one way or the other. Fingers are crossed, and it's
hoped that this will have been the better move.

Properties are still around though, but will be gone soon.