Add class 'method'. @main is now a method.
So begins the start of splitting lily functions and C functions. This is done
in hopes that it will simplify things later on. It will allow functions to be
smaller (the signature will not include an allocation for code), and methods
won't have to have a pointer for the c func.

It also makes it so functions can't be assigned to methods and vice-versa.
This allows the vm to know for sure if it's calling a method or a function,
instead of guessing each time (a small speedup). Functions may be set to be
unchangable later too.