Reduce lexer buffers from 1028 to 128.
This saves around 2.4 KB of memory on the test runs. This is over half for
hello_world.ly, and a bit under that for basic_types.ly.

It's thought that most users won't have lines exceeding 128 chars in size.
This may prove to be untrue later if, for example, the server compresses the
page before parsing. Maybe. I'll get to that when it comes.