read_line validates utf-8. String test added to basic_types.ly.
read_line currently does lily_raise in the event that there's bad utf-8,
rather than writing the utf-8 invalid character. It was much easier to do, and
much less chance of an error. It will also allow identifier scanning to be
simpler.

It can easily be fixed later on, if it's an issue.