A basic if test now works. Fixed vm macro cases.
The macros were grabbing the wrong areas (comparison), and flat out forgetting
a case (arith). The macros had to be fixed, because the condition test relied
on them.

The new condition code is pretty fragile. It recognizes if, elif, and else,
but only parses an if with a basic condition. The jump will fail on a str, and
also doesn't validate the tree well (so a func call tree or a tree that
doesn't set the ast's result will cause a crash.