Add KEY_RETURN. Check for it in lily_parser.c. Fix target/save in emitter.
So at this point, it kicks back an error about str+str not being implemented.
But this is another step toward return statements working.

The emitter now saves the var, and has target return type, code, and var. This
allows the parser easy access to what it wants (the info for the current var),
with the save/restore needed to implement methods.