emitter: Fix oops in check for making new storages.
Storages are allocated in a circular fashion, and the id goes up with each
expression. If the next id is the SAME as the current expression id, then this
expression has already claimed it (unsafe). If it is DIFFERENT, then some other
expression is using it (safe).

I had it accidentally backwards.