Fix multi-line else requiring a word after 'else'. Tests added.
Now, expression_value validates the token that is sent to it for being a val
value. This is a better way of doing things anyway, because it allows for
expressions that don't start with a word. For example:

I also made the token after ) be scanned as a binary op, since it should be
a binary op. This was blocking the method.ly test from passing (the @> was
being checked for being a value, and not coming out right).

if "a" == var
if 10 > var

and so on are now okay in multi-line expressions.
Two tests added: One to sanity.ly, and one to method_test.ly.