Dump the correct vars when scoping out. Two tests added...
The problem always occured on the third method, since the code to be executed
for the second and all subsequent var closures was very wrong.

Another test was added to check that large resizes for calls are working
properly. This test wasn't added before because I ran into the above bug when
trying to add it.