ast: Save root with the call tree on entry. Fixed bad reparenting. Test added.
Previously, the root of the ast tree was treated as THE top tree, above all
others. This seemed to work well, except when...

add(1 + 2 > 3 + 4, 5)

would result in the first addition not even showing up. This was because there
was no tree at the top of the expression, only the current. So when the current
was fixed, the 1 + was lost.

Now, root is saved for every call entry, and is used instead of current where
appropriate.

merge_tree was fixed so that left and right are swapped if the right side has
a higher priority. Also, it uses tree instead of active, which is right.