ast: Drop the 'args_collected' part of asts. Just do a direct grab as-needed.
Only one place in the code uses this. So now that code will just take the
result and check it in that function, instead of saving to the ast. This saves
about 32 B of memory, which is something. There shouldn't be a performance
loss, since the code is just executing at a different place, nothing more.