The concat function works...sort of.
It could probably use a more efficient algorithm though, along with a bunch of
other stuff.

The vm function call op was fixed so that the return would be arg[0], meaning
the first arg is at [1], the second at [2], and so on. It seems easier to do
that than to put the arg at the end (first at [0], second at [1]...).

str storages are set to a NULL value on creation, since concat checks for
previous ones and lily_free_symtab will want to make sure all the strvals
allocated are deleted. There might be memory leaks resulting from this, but
there's no gc yet to fix that.