Strings now have a refcount, which fixes some double-free issues. 1st step.
I recently found out that when sanity.ly was causing some double frees because
both the storage and the literal had the same copy of data and were trying to
destroy it. o_str_assign was added so that strings would try to ref and deref
string data appropriately, and it's a start.

A few things don't work right now, like strings assigning to objects, and
strings that are sent and returned result in some extra frees...I think. More
tests are to come.