The emitter writes line info after the opcode, except for jumps and returns.
A lot of magic numbers were bumped forward one spot to accomodate. Tested this
with lily_strload, sanity, and hello_world. This was considered to be simpler
than creating an additional table that would map opcodes to line numbers. It's
simpler (I think), because we don't have 2x space from storing a line number
in one table, and a line number. Even with a compressed version, where the
spots would represent maximums (ex: Opcodes >= opcode [i] map to line [i]),
the space saved would just complicate stuff.

If it's found that the table way of storing line information will be better,
then that will be used in the future. So if you're reading this in the future,
consider this open to better ideas.

This has been on my todo list for months now, so I'm glad to FINALLY see it
get done. Now code that fails can have a line number to reference.