emitter: If a call needs an object, 'cast' as needed. Sanity test added.
Previously, a call that took an object would require objects be passed to it.
This was silly, because anything can be an object, and objects can already be
assigned non-object values.
Now, if a call needs an object, it will do an object assign to a storage, then
do the call.
This will be used later for fprintf-style calls.