emitter: Do type-checking for the varargs of a callable.
Each of the extra arguments is checked against the last sig of the vararg
callable. This allows a caller to specify a vararg function that, say, takes
only 1+ strings, instead of 1+ anything and checking for strings. Or a vararg
callable that takes a string, and 1+ objects. (ex: an fprintf-style call).