Add stub function printfmt(str, object...). Set VAR_SYM flag for builtins.
For some reason, there's an object assignment resulting from the printfmt
call now in sanity.ly. Somehow this results in an object copy, which targets
a builtin, so VAR_SYM needed to be set for builtins. It should have been set
already.

This is probably incorrect, but it's a step forward.