Send debug to stderr, html to stdout. Fixed a couple html bugs in lexer.
One bug was that @> was only skipping @, so the html would start with the > of
the lily closing tag.
The other was that the page scanner was using the wrong position to terminate
when <@lily was found (lbp is the position within the current line, htmlp is
the position in the html buffer.)

With this commit, it's now possible to redirect the output of a Lily script
to some file to see if it works as intended. This is the start of useful
printing of data in Lily!