vm: Fix various memory issues in grow_vm and new state alloc. Many oopses.
In allocating a new state, the interpreter isn't setup so lily_raise_* must
not be used. Instead, null must be returned so the parser can free everything
and die gracefully.
In growing the vm...use lily_realloc when growing, not lily_malloc. I'm going
to pretend that I never committed that.