Fix 'integer a = 1, b = 2' not working aka decl list initialization broken.
The expression handler was catching the ',' and thought that it was part of
a call, resulting in valgrind warnings and a crash.

Added a sanity test which also checks that a decl list can handle a call
before the comma (because it should be able to), to allow for
integer a = add(1, 2), b = 4... kinds of assignments.