Recursion, part 1: Implement saving of locals in emitter, printing in debug.
To do this, offsets were added to each method so that it's kind of easy to
figure out how many vars need to be saved. Offsets start at the id of the
method, and are added to each time variables are moved into the symtab's
old_var_table. Current id - offset = number of locals to save starting from
after the method.

There are a few caveats to this:
* Storages are not saved.
* In a declaration expression (ex: integer a = add(1,2)), a is saved.
* The vm doesn't actually perform the save.