Emitter: Make walk_tree's call case a bit more bearable (rewrite).
In particular, limit the number of WRITE_PREP_LARGE calls to 1, and work the
rest of the code around that. It should be a tiny bit faster, but that's more
of a guess. Also, it uses a memcpy for copying the save cache to code, instead
of a loop. So that should be faster.