Milestone 1: Implement o_save, o_restore. Test added to save_test.ly.
Added a save stack to vm to get this to work. It's a bit rough, but what's
important is that save and restore now finally, finally work. Fibonacci is
implemented a bit strange, because of a couple quirks that I'll fix pretty
soon.

This feels like a huge accomplishment, and the start of something awesome.