Emitter: Save lhs if a method's parent is the rhs of a binary. Test changed.
Since the binary expression isn't complete yet, the lhs may be modified by the
call included in the rhs. So it has to be saved as well. This resulted in the
fib method in save_test.ly having to run fib(n-1) and fib(n-2) on different
lines, instead of as a single expression.