Parser: Allow return in oneline if, drop word checks. Sanity test added.
This was causing the fib test in save_test.ly to need to be written strange.
I also noticed that oneline if's were not allowing the condition to be a
literal of any sort (disallowing things like if "a" == somevar), which was
inconsistent with what multiline if allows. So I fixed that too.

12 tests now, all passing.