Using AFT, found a bunch of potential memory-related problems.
In particular, the idiom used when reallocing could result in memory not being
free'd. The parser needs to create new symbols before creating parts that go
in the symbols, since new sym creation expects to succeed. There was one case
of realloc directly being used (lily_pkg_str, and this took about an hour to
figure out).

This was discovered by running the aft test from 0 to 846 on sanity.ly. 846
is the total number of allocs in sanity.ly so far. This patch is a bit large,
but there shouldn't be many more big patches like this. I intend to run AFT
at least once every few builds to make sure no new issues creep up.