emitter: Walk trees more selectively when a binary op. Catch binary = ...
The lhs and rhs trees were being walked regardless of op. This is wrong,
because assign requires a var at the left side (and thus should never be able
to walk that side). It now checks for a non-var left side and produces an
error message for that case. The check has not been tested yet.