and / or are finally in now. sanity.ly updated with a lot of tests.
This took way longer than it should have. Hopefully, the verbosity of the
tests in sanity.ly will ensure that and/or don't break.

The current version of each yields a 0 or 1 value as a result, instead of
possibly yielding one of the values. This may or may not change in the future.