str package: Don't free the str if the realloc fails!
Line 510 of sanity.ly reads:
    if (s.concat("a")).concat("a") != "aaa":

This was causing the realloc case in lily_pkg_str's lily_str_concat to be hit,
which was free'ing the strval and the string...without dropping the refcount.
Later, it would get free'd again. If the realloc fails, don't free anything,
or the refs will be off.

What makes me happy is that I discovered this by accident from running the AFT
tests (0..1061 now, I think), and with the limit set to 1059 I hit the warning
about an invalid free and a double free.

In short, aft is awesome.