emitter: Re-organize sig compares for calls, and fix bug #4.
If a function took an object and was given it...the object would be
o_obj_assign'd into...another object. So sending objects to printfmt would
fail because it looked inside of the object and found another one.

This re-ordering also sets a nice stage for later things like comparing parts
of a func/method, or more flexible casting.