lily_new_var -> lily_try_new_var. Unify method+func creation.
Previously, builtin functions were registered in the symtab in a way that they
would not copy the name, and much of the code was copy+pasted from
lily_new_var, with some slight altering.

Now, functions are created through lily_try_new_var. This is a step toward
functions being parameters to a method.

@main is now registered separately from the builtin functions, and thus no
longer has a 'seed' in lily_seed_symtab.h.