Parser can pass calls as args now. vm grabs method_val each time for calls.
The current functions are not passable, because the language has no way to
convey var args yet. So they would fail the sig compare (when that gets
fixed).

Previously, the emitter was writing the method's method_val/func's actual
func when writing a call. Now, the vm gets the var and has to deal with the
var. The old way was causing crashes for methods as args, because the emitter
would write in the invalid method_val for the method arg var. The initial,
invalid method_val. Now it has to grab from the given var each time.

This commit exposes a new bug, wherein the method_val structs will need
refcounts or...something because there are a lot of invalid free messages.