list[int] x finally works! Hurray! list_test.ly temporarily added.
As usual, test_complex_decl.ly and list_test.ly will be merged back into
sanity.ly once the code is more stable. Lists are able to be declared now, but
are very rough. As with complex decls, a lot of unfinished features and/or
bugs are ahead.

list_test.ly shows that lists don't show their inner type. The bugs will be
handled one at a time, and eventually both tests will be merged back into
sanity.ly together.

I've wanted lists for about a year and a half, so I consider this a major
advancement. It'll be even better when they work proper!