Make lily_ast_enter_call more generic (call->subexpr). List decl test added.
There's a lot of common requirements between calls, parenth expressions, and
list expressions. So it made sense to make lily_ast_enter_call into
lily_ast_enter_subexpr, as well as making it more generic to allow for
receiving lists.

This is only part of making list expressions work. Note that the test in
list_test.ly raises an error, instead of working. More to come.