Added 'str' and 'method' to lily_value, strval->str_val. Cleanup.
A lot of code was using ptr and casting it. This is no longer necessary. Also,
lily_strval was changed to lily_str_val to be consistent with lily_method_val.
A lot of cleanup was done to get rid of the now-unnecessary use of .ptr.

Functions still use .ptr, but that will be fixed later. Their use is so little
that it's not important (yet?).