At long last, building lists inside of lists. Test added too.
Sigcomp within emitter had to be updated to support lists.
Symtab now won't do a list sig free for storages. This allows storages to
freely reuse signature information without crashing. So only the initial
sig of a list can be deleted for storages. This will require that storages
be deleted first later on, but that's okay.
Fixed building lists using the storage directly, instead of the value.
A few other misc things too.

There are still some issues with free-ing all of the list data at exit.
However, this is a big step forward. This leaves only some minor list
cleanups, and subscripts.