Signatures, even simple ones, now have refcounts.
Since o_build_list came in, storages have been using the signature of the
first element. This was done so that storages would have a complete signature
when doing emitting. It couldn't be malloc'd, or it could be leaked. However,
this resulted in certain signatures being leaked.

This caused a problem when, say, a list[integer] that a list[list[integer]]
depended on was free'd first. When it came time for the list[list[integer]]
to be destroyed...it didn't have an inner sig.

This will hopefully be useful later on for modules to complex signatures.

This is yet another hack to make lists work. Things will be touched up later.