Merge test_complex_decl.ly, list_test.ly into test_complex.ly. Cleanup.
Complex method decls and lists have a lot to do with each other. It makes
sense to put them together, so that they can be tested together. I plan to
extend the tests in test_complex.ly before finally merging it back into
sanity.ly. There are far too many things not covered and not yet working to
merge it back to sanity.ly