Emitter: Better bad assignment info. Also, fully print lists now.
Previously, trying to do list[integer] to list[str] would result in:

'Cannot assign list to list.'

Now, it prints the correct message:

'Cannot assign list[str] to list[integer].'

Not only was list info not being printed, but they were printed in the order
of left, then right. So what was printed was backward anyway.