BIG: Redo collect_args from the ground-up, into collect_var_sig.
It was determined that collect_args was designed in a way that made things
problematic. The biggest problem was that it allowed creating either vars or
sigs, and pulling sig data from vars or the sig stack. So if you had a method,
it had to do a giant check to see where it would get the args from.

Now, collect_var_sig just stuffs everything into the sig_stack, and everything
pulls from there. It's got helper functions, instead of being several pages
long.

lily_try_new_var no longer takes a class. It now takes sigs, but does not ref
them. This is because a signature is often created and directly handed to a
var. So this would result in an unnecessary ref/deref for a lot of areas.

init_sym_common no longer creates a sig and sets the value. It now does only
the value part, and will probably be changed later.

lily_sig_for_class was made to handle sig creation. test_complex.ly was
updated with a bunch of complex declaration tests, and the old checking code
was unceremoniously removed. So far, everything works with complex sig
creation.

Lists and methods can now be inside of each other (though lists are still a
bit useless at this point). This code is expected to be buggy, and aft is
expected to crash. That will be fixed later.