Object values now have a refcount, sig, and value.
Previously, objects were very strange. The sig of the object's value was
stored in the sig of the sym holding the object. For a long time, this was
fine.

Then lists came along. It wasn't possible to have a list of objects because
each value did not have a sig. Oops. So with this, it's possible to create a
real list of objects and know what's inside of each of those objects (to
protect against info loss).