Subscript assign works...and opens bucket of worms too.
With this patch, it's now possible to do lhs[index] = rhs. Being able to
change the value of a list is, obviously, an important feature. However, lists
also support objects, which can hold any kind of item inside of them. So this
brings out a couple of bugs that will be fixed in the future:

It's now possible to have a list of objects that has itself as an element.
This results in the parent list having too many references, and thus never
being deleted. This has been confirmed.
If two lists have an element that refers to the other, there's a chance for
either an infinite deref loop, or a memory leak. This has not been verified.