At long	last, merge test_complex.ly into sanity.ly...with caveats.
At this	point, the list	and complex signature code is stable enough to finally
merge it into sanity.ly. There are a few parts that are	still incomplete,
though that is due to missing features elsewhere.

Some caveats:
* Object subscript assign can't be tested because objects can't be cast to
  other types.
* Function subscript assign can't be tested because all	functions have
  different signatures.
* Method subscript assign can't	be tested because the result of	a subscript
  cannot be called.
* It's possible	for list[object] to have a value that references itself,
  which results in the list never getting collected.

I've decided to merge test_complex.ly in because, overall, the complex sig and
subscript stuff is finally stable now. Aside from the above, aft reports 0
memory leaks.