Emitter: sigequal handles callables now. Failtest added. Fixes #21.
Previously sigequal would not handle callables, instead just assuming they
matched. When emitting a return, the value is checked to be a more exact match
to what the method should return. Also fixed where a list of methods would
accept the wrong kind of method.

A test was added to lily_failtest to check that lists of method won't accept
the wrong kind of method.