nil really means nil now. Lists store nil-ness. Fixed an aft bug, and #26.
Previously, some things were more nil than others (book reference!).
free_sym_common would just blast everything at the end and hope for the best.
It worked...sort of. Things that were supposed to be nil were getting free'd.

With this fix, nil really means nil. If symtab creates an initial value, then
it marks that as being not nil. free_sym_common no longer blasts sym values
without checking nil-ness first. Lists also store nil-ness of their args
per-element (this may be changed later, as it could be prohibitively expensive
for large lists, and may/may not have any usefulness).

If a block was reused by AFT, then every free would seem to be of that block.
It was kind of funny seeing a thousand free's of one really really sturdy
block.

The str pkg was also setting the str return to not nil if it failed to
realloc. This is a no-no, and was caught by aft+sanity.ly.

Added a test in to check for future occurances of #26.