Emitter now caches storages with complex sigs. Lots of simplification.
Previously, emitter would grab any available list storage when doing either
subscript or build list. It would then ref/deref the inner element of the
storage. When doing these ops...the vm did the same thing.

Now, the emitter caches any storage with a complex sig (currently, lists and
calls). This sig is reused when possible, so that the inner sig does not have
to be ref/deref'd.

This was done because it saves emitter a lot of needless work in ref/derefing
list signatures. Since the VM runs many times and the emitter once, offloading
this work to an emitter cache was considered a good tradeoff. This also helps
reduce extremely annoying forgotten ref/deref issues.