lily_strload -> lily_cliexec. It now runs argv[1].
This was originally done a while back while doing some extensive ast tests. I
decided to merge it in, because strload was useless as it was. lily_failtest
verifies that string-based loading works, as well as doing useful tests.
lily_cliexec is intended to be used for quick code checks without having to
modify a file.