Create lily_syminfo.h to hold basic sig+value info. msgbuf allows sigs now.
For lily_msgbuf to support adding signatures to the buffer, a lot of data from
lily_symtab.h was pulled into a new file named lily_syminfo.h. The code for
adding a sig to the msgbuf was pulled from the emitter and put into
lily_msgbuf so it can be shared.

This commit is a step toward allowing typecasts.