Objects can FINALLY be typecast to other things. Fixes bug 17.
This has been a long time coming. With this commit, objects or anything that
returns an object can finally be typecast to another type. This makes objects
useful. The integer in an object can finally be extracted and assigned to
something else. Or used.

The syntax is a bit different than C. It goes like:

@(type:value)

This makes typecasts easy to disguish from parenth expressions. The syntax
leaves no room for ambiguity (ex: Can't accidentally cast only one variable
instead of the whole expression). Number to/from integer will be handled in a
future commit.

As of now, only object can be cast from. No other classes could work with
casting to something else, except for integer<->number (expect that in a
future commit). So it makes sense, I think. But objects can be cast with the
same flexibility as argument declaration (cast to list of method/method taking
whatever params, etc).

Updated sanity.ly with a new test.