vm: Subscript needs to take objects into account.
Previously, subscripts would just assign a value for anything that was
refcounted and bump the value's refcount. This was incorrect for objects.
Objects never get refcounted unless they are copied to somewhere else (ex:
saving a value, or putting an object into a list). So subscript was causing
objects to leak because it was doing a simple assign which overwrote the
container.

Example:

object o1 = 10
list[object] o2 = [o1]
object o3 = o2[0]

This caused a leak of a single object value, because subscript was writing
over the storage's initial object value.

Subscript now does the right thing for objects.