Fix decl not allowed at end of method.
This is useless to allow, because whatever was declared is unnecessary.
However, it's even worse to raise a syntax error for it. The interpreter will
probably issue a warning in the future.

I've run into this a few times when doing some tests where I just wanted to
check a declaration.