Fix: !intlist[0] now works. Also fixed unary's line num. Sanity updated.
It turns out that unary merge was not doing an oo_merge for subscripts coming
in. This was causing...some magic that resulted in the emitter seeing
!intlist, and then rightly complaining about it.

Additionally, I discovered that unary trees didn't set a line number. This
would have caused some serious weirdness if a unary op was the last op to
fail. Finally, fixed 'merge_oo' saying it was 'merge_unary' in the comment
over it. This was an accident from when I redid comments a while back.

Added a test to sanity's unary checker.