Fix: Subscript of list[object] may leak, depending on the value.
list[object] o1 = ["1", "2", "3", "4"]
list[object] o2 = [o1[0]]

This was causing a problem but something like this:

list[object] o1 = ["1", 1.1, 10]
list[object] o2 = [o1[0]]

was not.

It turns out that object subscripts were not doing a proper deref of the
result before putting a new value in. The lily_deref_unknown val code was
called on them, which ignored the object that was passed. This was also
incorrect, because the value in the object should be deref'd instead of the
object itself.

I also modified the subscript code to select for !object first, instead of
checking for refcounting and then object. This is assumed to be a bit faster,
but may not be.