vm: List build now copies the values of objects given into new objects.
This was done because not doing it resulted in objects acting like pointers
when it made no sense for them to be doing so. For example:

o = [10]
o = [o]

Before this patch, o would become a circular list consisting of itself. This
is odd, because the object previously held a list of 10.

After this patch, o would be [[10]], which makes sense since o was [10], then
put inside of another list.

What this patch does to make o_build_list make new objects to hold the values
of objects given to it. This keeps objects being circularly ref'd when they
shouldn't be.

This isn't a fix for #24, but it addresses a way that objects could be
incorrectly circularly referenced.