Big circular referencing leak fix. Fixes bug #24.
The val_is_nil part of lily_list_val struct is now flags, and an int as well.
It holds S_IS_NIL to correspond to a nil value inside, and the new
S_IS_CIRCULAR to indicate that the element contains a circular reference.

Added a pretty big circular ref test to sanity.ly, which is working rather
nicely.

circle_buster is responsible for fixing up circular references, as well as
making lists mark their elements as S_IS_CIRCULAR.